+++ my-canister-dapp-rs/my-canister-dashboard/CHANGELOG.md
@@ -2,6 +2,18 @@
 
 ## [Unreleased]
 
+## [0.10.2] - 2025-08-19
+
+### Fixed
+
+- Canister dashboard frontend: improved Top Up UX
+
+## [0.10.1] - 2025-08-16
+
+### Changed
+
+- Version bumps, new build files
+
 ## [0.10.0] - 2025-08-13
 
 ### Added